feat: isolated mapper, added tests
