+++ configuration/src/test/java/io/camunda/configuration/SecondaryStorageElasticsearchTest.java
@@ -49,6 +49,12 @@ public class SecondaryStorageElasticsearchTest {
 
   private static final int EXPECTED_NUMBER_OF_SHARDS = 3;
 
+  private static ExporterConfiguration fromArgs(final Map<String, Object> args) {
+    return new io.camunda.zeebe.broker.exporter.context.ExporterConfiguration(
+            "camundaexporter", args)
+        .instantiate(ExporterConfiguration.class);
+  }
+
   private static final boolean EXPECTED_HISTORY_PROCESS_INSTANCE_ENABLED = false;
 
   @Nested
@@ -123,8 +129,7 @@ public class SecondaryStorageElasticsearchTest {
       final Map<String, Object> args = camundaExporter.getArgs();
       assertThat(args).isNotNull();
 
-      final ExporterConfiguration exporterConfiguration =
-          UnifiedConfigurationHelper.argsToCamundaExporterConfiguration(args);
+      final ExporterConfiguration exporterConfiguration = fromArgs(args);
       assertThat(exporterConfiguration.getConnect().getType()).isEqualTo(expectedType);
       assertThat(exporterConfiguration.getConnect().getUrl()).isEqualTo(expectedUrl);
       assertThat(exporterConfiguration.getConnect().getUsername()).isEqualTo(EXPECTED_USERNAME);
@@ -260,8 +265,7 @@ public class SecondaryStorageElasticsearchTest {
       final Map<String, Object> args = camundaExporter.getArgs();
       assertThat(args).isNotNull();
 
-      final ExporterConfiguration exporterConfiguration =
-          UnifiedConfigurationHelper.argsToCamundaExporterConfiguration(args);
+      final ExporterConfiguration exporterConfiguration = fromArgs(args);
       assertThat(exporterConfiguration.getConnect().getType()).isEqualTo(expectedType);
       assertThat(exporterConfiguration.getConnect().getUrl()).isEqualTo(expectedUrl);
       assertThat(exporterConfiguration.getConnect().getUsername()).isEqualTo(EXPECTED_USERNAME);
@@ -373,8 +377,8 @@ public class SecondaryStorageElasticsearchTest {
       final Map<String, Object> args = camundaExporter.getArgs();
       assertThat(args).isNotNull();
 
-      final ExporterConfiguration exporterConfiguration =
-          UnifiedConfigurationHelper.argsToCamundaExporterConfiguration(args);
+      final ExporterConfiguration exporterConfiguration = fromArgs(args);
+
       assertThat(exporterConfiguration.getConnect().getUrl()).isEqualTo("http://wanted-url:4321");
     }
   }
